+++ flutter_inappwebview/lib/src/in_app_browser/in_app_browser.dart
@@ -378,10 +378,7 @@ class InAppBrowser implements PlatformInAppBrowserEvents {
   void onPermissionRequestCanceled(PermissionRequest permissionRequest) {}
 
   @override
-  FutureOr<bool?>? onPrintRequest(
-    WebUri? url,
-    PlatformPrintJobController? printJobController,
-  ) {
+  FutureOr<bool?>? onPrintRequest(WebUri? url) {
     return null;
   }