+++ flutter_inappwebview/lib/src/in_app_webview/headless_in_app_webview.dart
@@ -133,11 +133,7 @@ class HeadlessInAppWebView {
       InAppWebViewHitTestResult hitTestResult,
     )?
     onLongPressHitTestResult,
-    FutureOr<bool?> Function(
-      InAppWebViewController controller,
-      WebUri? url,
-      PlatformPrintJobController? printJobController,
-    )?
+    FutureOr<bool?> Function(InAppWebViewController controller, WebUri? url)?
     onPrintRequest,
     void Function(InAppWebViewController controller, int progress)?
     onProgressChanged,
@@ -408,8 +404,7 @@ class HeadlessInAppWebView {
                      onUpdateVisitedHistory.call(controller, url, isReload)
                : null,
            onPrintRequest: onPrintRequest != null
-               ? (controller, url, printJobController) =>
-                     onPrintRequest.call(controller, url, printJobController)
+               ? (controller, url) => onPrintRequest.call(controller, url)
                : null,
            onLongPressHitTestResult: onLongPressHitTestResult != null
                ? (controller, hitTestResult) =>