+++ flutter_inappwebview_ios/lib/src/in_app_webview/in_app_webview_controller.dart
@@ -1073,25 +1073,15 @@ class IOSInAppWebViewController extends PlatformInAppWebViewController
         if ((webviewParams != null && webviewParams!.onPrintRequest != null) ||
             _inAppBrowserEventHandler != null) {
           String? url = call.arguments["url"];
-          String? printJobId = call.arguments["printJobId"];
           WebUri? uri = url != null ? WebUri(url) : null;
-          IOSPrintJobController? printJob = printJobId != null
-              ? IOSPrintJobController(
-                  IOSPrintJobControllerCreationParams(id: printJobId),
-                )
-              : null;
 
           if (webviewParams != null) {
             return await webviewParams!.onPrintRequest!(
               _controllerFromPlatform,
               uri,
-              printJob,
             );
           } else {
-            return await _inAppBrowserEventHandler!.onPrintRequest(
-              uri,
-              printJob,
-            );
+            return await _inAppBrowserEventHandler!.onPrintRequest(uri);
           }
         }
         break;