+++ flutter_inappwebview_platform_interface/lib/src/in_app_webview/platform_webview.g.dart
@@ -547,9 +547,6 @@ enum PlatformWebViewCreationParamsProperty {
   ///
   ///**Parameters - Officially Supported Platforms/Implementations**:
   ///- [url]: all platforms
-  ///- [printJobController]:
-  ///    - Android WebView
-  ///    - iOS WKWebView
   ///
   ///Use the [PlatformWebViewCreationParams.isPropertySupported] method to check if this property is supported at runtime.
   ///{@endtemplate}